added point about format of data and additional (see desc.)
Under the "Contributing" heading I have added a disclaimer. This is to make the audience aware that the checklist is referencing the properties data should have irrespective of the system. Whereas the methods to clean the data will often be in reference to Power BI. The cleaning activities can be performed in other respective tools/software,

Also a typo has been fixed.